ThresholdControls: Use correct scale key when calculating handle position
